app/requirements.txt
apprise==1.9.0
requests==2.32.2
urllib3==2.2.3
PyYAML==6.0.2
python-telegram-bot==21.6
beautifulsoup4==4.10.0